tesseract/requirements.txt
boto3
botocore
pdf2image
pillow
pytesseract
streamlit